+++ backend/coins/btc/account_test.go
@@ -73,7 +73,6 @@ func TestAccount(t *testing.T) {
 				SigningConfigurations: signingConfigurations,
 			},
 			DBFolder:        dbFolder,
-			Keystore:        nil,
 			OnEvent:         func(accountsTypes.Event) {},
 			RateUpdater:     nil,
 			GetNotifier:     func(signing.Configurations) accounts.Notifier { return nil },