+++ backend/coins/eth/account.go
@@ -17,12 +17,14 @@ package eth
 
 import (
 	"context"
+	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"math/big"
 	"net/http"
 	"os"
 	"path"
+	"strconv"
 	"strings"
 	"time"
 
@@ -42,6 +44,7 @@ import (
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/sirupsen/logrus"
 )
 
@@ -53,6 +56,18 @@ const ethGasStationAPIKey = "3a61bee56f554cc14db4f49e2ace053b3086d3c323aefec83e5
 
 func isMixedCase(s string) bool {
 	return strings.ToLower(s) != s && strings.ToUpper(s) != s
+}
+
+// IsValidEthAddress checks if a string is a valid 20 byte ETH address and validates checksum when present.
+func IsValidEthAddress(addr string) bool {
+	if !ethcommon.IsHexAddress(addr) {
+		return false
+	}
+	// Validate checksum if the address is mixed case, see https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md
+	if isMixedCase(addr) && addr != ethcommon.HexToAddress(addr).Hex() {
+		return false
+	}
+	return true
 }
 
 // Account is an Ethereum account, with one address.
@@ -487,14 +502,10 @@ type TxProposal struct {
 }
 
 func (account *Account) newTx(args *accounts.TxProposalArgs) (*TxProposal, error) {
-	if !ethcommon.IsHexAddress(args.RecipientAddress) {
+	if !IsValidEthAddress(args.RecipientAddress) {
 		return nil, errp.WithStack(errors.ErrInvalidAddress)
 	}
 	address := ethcommon.HexToAddress(args.RecipientAddress)
-	// Validate checksum if the address is mixed case, see https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md
-	if isMixedCase(args.RecipientAddress) && args.RecipientAddress != address.Hex() {
-		return nil, errp.WithStack(errors.ErrInvalidAddress)
-	}
 
 	suggestedGasPrice, err := account.gasPrice(args)
 	if err != nil {
@@ -642,8 +653,13 @@ func (account *Account) SendTx() error {
 		return errp.New("No active tx proposal")
 	}
 
+	keystore, err := account.Config().ConnectKeystore()
+	if err != nil {
+		return err
+	}
+
 	account.log.Info("Signing and sending transaction")
-	if err := account.Config().Keystore.SignTransaction(txProposal); err != nil {
+	if err := keystore.SignTransaction(txProposal); err != nil {
 		return err
 	}
 	// By experience, at least with the Etherscan backend, this can succeed and still the
@@ -802,17 +818,211 @@ func (account *Account) VerifyAddress(addressID string) (bool, error) {
 	if !account.isInitialized() {
 		return false, errp.New("account must be initialized")
 	}
-	canVerifyAddress, _, err := account.Config().Keystore.CanVerifyAddress(account.Coin())
+	keystore, err := account.Config().ConnectKeystore()
+	if err != nil {
+		return false, err
+	}
+	canVerifyAddress, _, err := keystore.CanVerifyAddress(account.Coin())
 	if err != nil {
 		return false, err
 	}
 	if canVerifyAddress {
-		return true, account.Config().Keystore.VerifyAddress(account.signingConfiguration, account.Coin())
+		return true, keystore.VerifyAddress(account.signingConfiguration, account.Coin())
 	}
 	return false, nil
 }
 
 // CanVerifyAddresses implements accounts.Interface.
 func (account *Account) CanVerifyAddresses() (bool, bool, error) {
-	return account.Config().Keystore.CanVerifyAddress(account.Coin())
+	keystore, err := account.Config().ConnectKeystore()
+	if err != nil {
+		return false, false, err
+	}
+
+	return keystore.CanVerifyAddress(account.Coin())
+}
+
+// SignMsg is used for personal_sign and eth_sign messages in BBApp via WalletConnect.
+func (account *Account) SignMsg(
+	message string,
+) (string, error) {
+	bytesMessage, err := hex.DecodeString(strings.TrimPrefix(message, "0x"))
+	if err != nil {
+		return "", err
+	}
+
+	keystore, err := account.Config().ConnectKeystore()
+	if err != nil {
+		return "", err
+	}
+	signedMessage, err := keystore.SignETHMessage(bytesMessage, account.signingConfiguration.AbsoluteKeypath())
+	if err != nil {
+		return "", err
+	}
+	return "0x" + hex.EncodeToString(signedMessage), nil
+}
+
+// SignTypedMsg signs an Ethereum EIP-712 typed message in BBApp via WalletConnect.
+func (account *Account) SignTypedMsg(
+	chainId uint64,
+	data string,
+) (string, error) {
+	keystore, err := account.Config().ConnectKeystore()
+	if err != nil {
+		return "", err
+	}
+	signedMessage, err := keystore.SignETHTypedMessage(chainId, []byte(data), account.signingConfiguration.AbsoluteKeypath())
+	if err != nil {
+		return "", err
+	}
+	return "0x" + hex.EncodeToString(signedMessage), nil
+}
+
+// WalletConnectArgs are the tx proposal arguments received from Wallet Connect with Gas, GasPrice,
+// Value and Nonce being optional.
+type WalletConnectArgs struct {
+	From     string `json:"from"`
+	To       string `json:"to"`
+	Data     string `json:"data"`
+	Gas      string `json:"gas,omitempty"`
+	GasPrice string `json:"gasPrice,omitempty"`
+	Value    string `json:"value,omitempty"`
+	Nonce    string `json:"nonce,omitempty"`
+}
+
+// EthSignWalletConnectTx signs an Ethereum Tx received from WalletConnect.
+func (account *Account) EthSignWalletConnectTx(
+	// send: whether transaction should be broadcast after signing
+	send bool,
+	// chainId: allow specifying other IDs than 1 (ETH mainnet) for other EVM networks
+	// TODO L#940 we also need to connect to an appropriate RPC for each L2 network/sidechain
+	chainId uint64,
+	proposedTx WalletConnectArgs,
+) (string, string, error) {
+	var nonce uint64
+	var message ethereum.CallMsg
+	var gasPrice *big.Int
+	var value *big.Int
+
+	// Error if chaindId != account.coin.ChainID() (i.e. 1) until L2 RPCs and proper support are added
+	if chainId != account.coin.ChainID() {
+		return "", "", errp.New("Unsupported EVM Network. BBApp only supports Ethereum Mainnet at the moment.")
+	}
+
+	if !IsValidEthAddress(proposedTx.To) {
+		return "", "", errp.WithStack(errors.ErrInvalidAddress)
+	}
+	address := ethcommon.HexToAddress(proposedTx.To)
+
+	if proposedTx.Nonce != "" {
+		parsed, err := strconv.ParseUint(strings.TrimPrefix(proposedTx.Nonce, "0x"), 16, 64)
+		if err != nil {
+			return "", "", err
+		}
+		nonce = parsed
+	} else {
+		nonce = account.nextNonce
+	}
+
+	if proposedTx.Value != "" {
+		bigIntValue, ok := new(big.Int).SetString(strings.TrimPrefix(proposedTx.Value, "0x"), 16)
+		if !ok {
+			return "", "", errp.New("error setting transaction value")
+		}
+		value = bigIntValue
+	}
+
+	data, err := hex.DecodeString(strings.TrimPrefix(proposedTx.Data, "0x"))
+	if err != nil {
+		return "", "", err
+	}
+
+	message = ethereum.CallMsg{
+		From:     account.address.Address,
+		To:       &address,
+		Gas:      0,
+		GasPrice: big.NewInt(0),
+		Value:    value,
+		Data:     data,
+	}
+
+	gasLimit, err := account.coin.client.EstimateGas(context.TODO(), message)
+	if err != nil {
+		if strings.Contains(err.Error(), etherscan.ERC20GasErr) {
+			return "", "", errp.WithStack(errors.ErrInsufficientFunds)
+		}
+		account.log.WithError(err).Error("Could not estimate the gas limit.")
+		return "", "", errp.WithStack(errors.TxValidationError(err.Error()))
+	}
+
+	for _, t := range account.feeTargets() {
+		//TODO Let user choose gas price/priority
+		if t.code == accounts.FeeTargetCodeNormal {
+			if t.gasPrice.Cmp(big.NewInt(0)) <= 0 {
+				return "", "", errors.ErrFeeTooLow
+			}
+			gasPrice = t.gasPrice
+		}
+	}
+
+	tx := types.NewTransaction(nonce,
+		*message.To,
+		message.Value, gasLimit, gasPrice, message.Data)
+
+	keystore, err := account.Config().ConnectKeystore()
+	if err != nil {
+		return "", "", err
+	}
+	signature, err := keystore.SignETHWalletConnectTransaction(chainId, tx, account.signingConfiguration.AbsoluteKeypath())
+	if err != nil {
+		return "", "", err
+	}
+	//TODO edit signer to match chainID proposed by wallet connect
+	// account.coin.Net() will only incude ChainID 1 in its current *params.ChainConfig
+	// Needs to be set to the appropriuate chain id for each supported network
+	//TODO we also need to connect to an appropriate RPC for each L2 network/sidechain
+	signer := types.MakeSigner(account.coin.Net(), account.blockNumber)
+	signedTx, err := tx.WithSignature(signer, signature)
+	if err != nil {
+		return "", "", err
+	}
+	txHash := signedTx.Hash()
+	if send {
+		if err := account.coin.client.SendTransaction(context.TODO(), signedTx); err != nil {
+			return "", "", errp.WithStack(err)
+		}
+	}
+	rawTx, err := rlp.EncodeToBytes(signedTx)
+	if err != nil {
+		return "", "", err
+	}
+	return "0x" + hex.EncodeToString(txHash[:]), "0x" + hex.EncodeToString(rawTx), nil
+}
+
+// Address returns the account's single Ethereum address.
+func (account *Account) Address() (*Address, error) {
+	if !account.isInitialized() {
+		return nil, errp.New("account must be initialized")
+	}
+	return &account.address, nil
+}
+
+// IsERC20 checks whether an account is an ERC20 token account.
+func IsERC20(account *Account) bool {
+	return account.coin.erc20Token != nil
+}
+
+// MatchesAddress checks whether the provided address matches the account.
+func (account *Account) MatchesAddress(address string) (bool, error) {
+	if !IsValidEthAddress(address) {
+		return false, errp.WithStack(errors.ErrInvalidAddress)
+	}
+	accountAddress, err := account.Address()
+	if err != nil {
+		return false, errp.WithStack(err)
+	}
+	if ethcommon.HexToAddress(address).Hex() == accountAddress.Hex() {
+		return true, nil
+	}
+	return false, nil
 }