+++ frontends/web/src/components/appdownloadlink/appdownloadlink.tsx
@@ -21,10 +21,11 @@ import { Button } from '../forms';
 import { A } from '../anchor/anchor';
 
 export const downloadLinkByLanguage = () => {
-  const language = i18n.language;
-  switch (language) {
+  switch (i18n.resolvedLanguage) {
   case 'de':
     return URL_CONSTANTS.DOWNLOAD_LINK_DE;
+  case 'es':
+    return URL_CONSTANTS.DOWNLOAD_LINK_ES;
   default:
     return URL_CONSTANTS.DOWNLOAD_LINK_GLOBAL;
   }