+++ frontends/web/src/routes/account/account.tsx
@@ -240,7 +240,8 @@ export function Account({
   const actionButtonsProps = {
     code,
     canSend,
-    exchangeBuySupported
+    exchangeBuySupported,
+    account
   };
 
   return (
@@ -285,6 +286,7 @@ export function Account({
             </div>
             {isAccountEmpty && (
               <BuyReceiveCTA
+                account={account}
                 code={code}
                 exchangeBuySupported={exchangeBuySupported}
                 unit={balance.available.unit}