+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
 ## Unreleased
+
+## 4.40.0
+- Add support for watch-only - see your accounts and portfolio without connecting your BitBox02
+- Android: new setting to enable screen lock authentication when accessing the app
 - Drop support for SAI token
-- Log Javascript console messages in the app log.txt
+- Ability to connect Ethereum (mainnet) wallets to dapps using WalletConnect
+- Log JavaScript console messages in the app log.txt
 - Add amounts hiding feature to enhance privacy when using the app in public
 
 ## 4.39.0